[back][Modify]: schema 변경 user 와 게시글 연동

+++ back/src/routers/recruitRouter.js
@@ -14,13 +14,16 @@ recruitRouter.post('/recruits', loginRequired, async (req, res, next) => {
      #swagger.security = [{ "bearerAuth": [] }]
     */
 
-    const user_id = req.currentUserId;
+    const userId = req.currentUserId;
 
-    const { name, title, content, hashtag } = req.body;
+    const { title, content, hashtag } = req.body;
+
+    if (!title || !content) {
+      throw new Error('제목과 내용을 입력해 주세요');
+    }
 
     const newPost = await recruitService.addPost({
-      user_id,
-      name,
+      userId,
       title,
       content,
       hashtag,
@@ -65,6 +68,8 @@ recruitRouter.put('/recruits/:id', loginRequired, async (req, res, next) => {
      #swagger.description = '팀원 모집 게시글을 수정한다.' 
      #swagger.security = [{ "bearerAuth": [] }]
     */
+
+    const userId = req.currentUserId;
     const post_id = req.params.id;
     const title = req.body.title ?? null;
     const content = req.body.content ?? null;
@@ -73,7 +78,7 @@ recruitRouter.put('/recruits/:id', loginRequired, async (req, res, next) => {
 
     const toUpdate = { title, content, status, hashtag };
 
-    const updatedPost = await recruitService.setPost({ post_id, toUpdate });
+    const updatedPost = await recruitService.setPost({ userId, post_id, toUpdate });
 
     if (updatedPost.errorMessage) {
       throw new Error(updatedPost.errorMessage);
@@ -85,16 +90,16 @@ recruitRouter.put('/recruits/:id', loginRequired, async (req, res, next) => {
   }
 });
 
-recruitRouter.get('/recruitlist/:user_id', loginRequired, async (req, res, next) => {
+recruitRouter.get('/recruitlist/:userId', loginRequired, async (req, res, next) => {
   try {
     /*
-     #swagger.tags = ['recruit'] 
-     #swagger.summary = '특정 user의 게시글 확인하기' 
-     #swagger.description = '특정 user의 게시글 확인하기' 
+     #swagger.tags = ['recruit']
+     #swagger.summary = '특정 user의 게시글 확인하기'
+     #swagger.description = '특정 user의 게시글 확인하기'
      #swagger.security = [{ "bearerAuth": [] }]
     */
-    const user_id = req.params.user_id;
-    const posts = await recruitService.getUserPosts({ user_id });
+    const userId = req.params.userId;
+    const posts = await recruitService.getUserPosts({ userId });
     res.status(200).send(posts);
   } catch (error) {
     next(error);
@@ -104,9 +109,9 @@ recruitRouter.get('/recruitlist/:user_id', loginRequired, async (req, res, next)
 recruitRouter.get('/recruitlist', loginRequired, async (req, res, next) => {
   try {
     /*
-     #swagger.tags = ['recruit'] 
-     #swagger.summary = '게시글 목록' 
-     #swagger.description = '전체 게시글 목록' 
+     #swagger.tags = ['recruit']
+     #swagger.summary = '게시글 목록'
+     #swagger.description = '전체 게시글 목록'
      #swagger.security = [{ "bearerAuth": [] }]
     */
     const posts = await recruitService.getPosts();
@@ -124,8 +129,9 @@ recruitRouter.delete('/recruits/:id', loginRequired, async (req, res, next) => {
      #swagger.description = '팀원 모집 게시글 삭제한다.' 
      #swagger.security = [{ "bearerAuth": [] }]
     */
+    const userId = req.currentUserId;
     const post_id = req.params.id;
-    const deletedPost = await recruitService.deletePost({ post_id });
+    const deletedPost = await recruitService.deletePost({ userId, post_id });
 
     if (deletedPost.errorMessage) {
       throw new Error(deletedPost.errorMessage);
